% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_func.R
\name{get_an_airline}
\alias{get_an_airline}
\title{Get random airline(s)}
\usage{
get_an_airline(n = 1)
}
\arguments{
\item{n}{the number of airlines to get (defaults to 1)}
}
\value{
a dataframe with airlines from the DB
}
\description{
Get random airline(s)
}
