% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_func.R
\name{one_function}
\alias{one_function}
\title{One function}
\usage{
one_function(x)
}
\arguments{
\item{x}{the thing to be returned}
}
\value{
\code{x}, the input
}
\description{
A noop function that returns its only input.
}
